Added "schedule" command
Schedules are created useing "schedule" rather than "insert-into" because they
are different from data and exercises.  They are named based on when they are
created, to help with ordering and also because they are regularly modified.
They will be in a separate repository from the other data and handled quite
separately, so I think it makes sense to create them in a different manner too.