Added "SimpleFlashcard" model
A simple flashcard-based model to get started.  Data is expected to be a series
of fields defined each on a single line, with a colon separating the key and the
value.  Exercise metadata takes the form of free text with field keys contained
within square brackets.  Question and answer are separated by the string "---".

An example datafile:

    Question: What is the largest man-made lake in the world?
    Answer: Lake Kariba

Example exercise metadata:

    [Question]
    ---
    [Answer]

The user is shown the question and asked to enter the answer.  If their entry
matches precisely with the expected answer, they score full marks.  Otherwise,
they are shown the correct answer and asked to assess their own score.