Added "SuperMemo2" scheduler
This scheduler follows the SuperMemo 2 algorithm precisely, as defined
on the SuperMemo website at [1], with a sample implementation at [2].
There were two discrepancies between the text and the implementation:

1. It is not entirely clear from the text whether the E-Factor should be
   adjusted before or after the interval.  Since the E-Factor is used
   during calculation of the interval, this obviously impacts
   functionality.
2. The text reads "If interval is a fraction, round it up to the nearest
   integer," however the sample implementation follows rounding rules;
   i.e. it rounds up if the value is >= 0.5 but rounds down otherwise.

In both cases the behaviour of the implementation was favoured as its
implementation in Pascal leaves less room for ambiguity than the English
description.

One ambiguity which is not resolved by either the text or the sample
implementation is whether to adjust the E-Factor/interval in the case of
repeats (step 7 in the text).  I have chosen not to do this because I
think it would skew the results significantly; the revision period at
the end is for revision, not assessment.

Note that the SuperMemo 2 algorithm expects scores to range from 0-5,
whereas in our case the range of scores is defined by the model, and
normalised to 0.0-1.0.  The SuperMemo2 scheduler multiplies by 5 to give
scores in the range SuperMemo expects, which might not be optimal for
scores in a different range (such as those output from the
SimpleFlashcard model, which are in the range 0-3).

[1] http://www.supermemo.com/english/ol/sm2.htm
[2] http://www.supermemo.com/english/ol/sm2source.htm