reschedule: Update queue files after rescheduling
Since "srs queue" is quite an expensive operation, we want to try and
reduce calls to it as far as possible.  By the time we leave
"reschedule", we know whether the exercise should be included in the
current queue, so we might as well update the information then and
there.

Fixes #2 in github.