simple_flashcard: Go back to header fields
XML files are annoying to edit and prone to syntax errors.  They were
introduced in 313a29b (simpleflashcard: Use XML for card data), to get
around problems with multi-line data, but I was never happy with this
solution.

This commit reverts the changes made in 313a29b and replaces them with a
new system for handling multi-line data, which is closer to the old data
format.  It consists of a series of key-value pairs, separated by colons
as before.  A line is considered a continuation of the previous line's
data if either of the following conditions are met:

 * It doesn't contain a colon anywhere within the line
 * It begins with a colon or whitespace character

The following example demonstrates various multi-line cases:

    Key1: Value 1
    Key2: Value 2: Line 1
          Value 2: Line 2
    Key3: Value 3: Line 1
    :Value 3: Line 2
    Key4: Value 4, Line 1
    Value 4, Line 2

Leading whitespace is stripped, but not if it follows a colon.  Newlines
are stripped, but can be added manually using the familiar \n syntax.
If the first character is a colon to denote a line continuation, that
colon will be stripped, but any colons following it won't be.

Here's an example preserving whitespace for indentation reasons:

    Question: What does the following C code achieve?\n
    :#include <stdio.h>\n\n
    :void main()\n
    :{\n
    :	printf( "Hello, world!" );\n
    :}
    Answer: It prints "Hello, world!" to the console