Give schedules more descriptive names
We were using a timestamp; instead we combine the names of the exercise
and the data being scheduled in the form "[data].[exercise]" to generate
an easily-understandable name for each schedule.

next-new was originally relying on the filenames being a timestamp so it
could just do a simple sort on filename.  This has been updated to sort
based on file creation time instead, which should result in the same
ordering.